Fixed: Removed various 64 bit and OS X incompatibilities
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@6032 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/ogl_shader.h
@@ -37,7 +37,7 @@ class CShaderManager {
 		void Init (void);
 		void Destroy (bool bAll = true);
 		void Setup (void);
-		int Deploy (int nShader);
+		int64_t Deploy (int nShader);
 		int Alloc (int& nShader);
 		char* Load (const char* filename);
 		int Create (int nShader);
